Verification of the theme icons
Verification of the theme icons such as Account Upgrade, closed support tickets. Verification of the domain search engine. The list of icons has been organized in order in the images folder and a single css has been left integrating in the same page the css of the icons that was apart everything has been unified so as not to have to upload several files to your vpanel